Breaking API changes (meta / links)
This commit introduces a new concept of how to handle links / meta data
of data-objects in the included and embedded relationship pathes.

Store provides to flags 'addMeta' and 'addLinks', which will build a
nested 'links' and 'meta' attribute to a resolved model object.

If an object contains relationships with meta, the information will be
stored in the model.meta with the relationship name as an attribute.
If the model object already has meta data, the attributes +
relationship-name attributes will be mixed.

The same behavior goes for links.

If there is an included data object with a meta / links value
and a relationship data object with a different meta / links value, the
relationship data object values will be favored over the included part.

If the constructor flag 'throwWarning' is set, an error will be thrown
if a naming-collision of different meta / links attribute names occurs.
This options is disabled by default.