Handle relationships with missing data attribute
There was an issue causing an attribute for a relationship to be set to
an {}, since there a links attribute causes the entity to be
resolved.

Now, if there is a relationship provided with no data attribute, the
denormalized object sets the relationship attribute to null.

Example:

JSON-API:
{
    id: '1',
    type: 'event',
    relationships: {
        comments: {
            links: {...}
        }
    }
}

Denormalizes to:
{
    id: '1',
    type: 'event',
    comments: null
}